+++ CHANGELOG.md
@@ -2,8 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/) 
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [1.1.0] - 2018-05-11
+
+### Added
+
+- New function `regenerateId` to regenerate session IDs after a given interval
 
 ## [1.0.0] - 2018-01-27